requirements.txt
aenum==3.1.12
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
decorator==5.1.1
Django==4.1.7
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-extensions==3.2.1
django-redis==5.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
factory-boy==3.2.1
Faker==18.10.1
frozenlist==1.3.3
idna==3.4
ipython==8.12.0
isort==5.12.0
jedi==0.18.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
prompt-toolkit==3.0.38
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
xmltodict==0.13.0
yarl==1.8.2
